requirements.txt
psycopg2-binary==2.8.6
python-dateutil==2.8.1
requests==2.25.1
fastapi==0.65.1
uvicorn==0.13.4
retry==0.9.2
pydantic==1.8.2
